* Renamed calibration_types to calibration_key_types.
