+++ qiskit_experiments/calibration/calibrations.py
@@ -39,7 +39,7 @@ from qiskit_experiments.calibration.exceptions import CalibrationError
 from qiskit_experiments.calibration.parameter_value import ParameterValue
 from qiskit_experiments.experiment_data import ExperimentData
 from qiskit_experiments.calibration.calibration_extraction import CalibrationExtraction
-from qiskit_experiments.calibration.calibration_types import (
+from qiskit_experiments.calibration.calibration_key_types import (
     ParameterKey,
     ParameterValueType,
     ScheduleKey
